added logo, removed comments, added conditional for not sending empty message and added RandomUser script
