+++ public/manifest.json
@@ -1,6 +1,7 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "React Chat App",
+  "name": "Simple React Chat App created at Algebra for Frontend Dev React Course purposes",
+  "author" : "Melita Poropat",
   "icons": [
     {
       "src": "favicon.ico",
