+++ src/App.js
@@ -2,138 +2,7 @@ import React, { Component } from "react";
 import "./App.css";
 import Messages from "./Messages";
 import Input from "./Input";
-
-function randomName() {
-  const adjectives = [
-    "Ana",
-    "Anamarija",
-    "Branka",
-    "Bojana",
-    "Ante",
-    "Andrej",
-    "Bruno",
-    "Darko",
-    "Dajana",
-    "Franko",
-    "Floria",
-    "Goran",
-    "Hrvoje",
-    "Ivona",
-    "Ivana",
-    "Maja",
-    "Marko",
-    "Mia",
-    "Lucija",
-    "Nika",
-    "Ema",
-    "Sara",
-    "Marta",
-    "Eva",
-    "Petra",
-    "Elena",
-    "Marija",
-    "Rita",
-    "Mila",
-    "Lana",
-    "Klara",
-    "Laura",
-    "Luka",
-    "David",
-    "Petar",
-    "Jakov",
-    "Ivan",
-    "Matej",
-    "Leon",
-    "Mateo",
-    "Mihael",
-    "Filip",
-    "Borna",
-    "Karlo",
-    "Fran",
-    "Josip",
-    "Lovro",
-    "Noa",
-    "Niko",
-    "Gabriel",
-    "Vito",
-    "Tomas",
-    "Jan",
-    "Roko",
-    "Lara",
-    "Katja",
-    "Tena",
-    "Lena",
-    "Maša",
-    "Una",
-    "Vita",
-    "Sofia",
-    "Korina",
-    "Tin",
-  ];
-  const nouns = [
-    " Voda",
-    " Rijeka",
-    " Lipa",
-    " Mjesec",
-    " Kiša",
-    " Vjetar",
-    " More",
-    " Jutro",
-    " Snijeg",
-    " Jezero",
-    " Sunce",
-    " Bor",
-    " Sjena",
-    " List",
-    " Zora",
-    " Sjaj",
-    " Šuma",
-    " Brdo",
-    " Planina",
-    " Oblak",
-    " Ptica",
-    " Konj",
-    " Knjiga",
-    " Lisac",
-    " Leptir",
-    " Grm",
-    " Prašina",
-    " Polje",
-    " Vatra",
-    " Cvijet",
-    " Ruža",
-    " Konjic",
-    " Lav",
-    " Trava",
-    " Munja",
-    " Dan",
-    " Noć",
-    " Tama",
-    " Pahuljica",
-    " Tišina",
-    " Glas",
-    " Glazba",
-    " Nebo",
-    " Oblik",
-    " Lotus",
-    " Drvo",
-    " Pero",
-    " Stablo",
-    " Tigar",
-    " Pjetao",
-    " Zec",
-    " Trešnja",
-    " Dunja",
-    " San",
-    " Žaba",
-    " Magla",
-    " Dim",
-    " Zvijezda",
-  ];
-  const adjective = adjectives[Math.floor(Math.random() * adjectives.length)];
-  const noun = nouns[Math.floor(Math.random() * nouns.length)];
-  return adjective + noun;
-}
+import randomName from "./RandomUser";
 
 function randomColor() {
   return "#" + Math.floor(Math.random() * 0xffffff).toString(16);
@@ -173,7 +42,7 @@ class App extends Component {
     return (
       <div className="App">
         <div className="App-header">
-          <h1>Chat App</h1>
+          <h1>React Chat @ Algebra</h1>
         </div>
         <Messages
           messages={this.state.messages}
