+++ src/Input.js
@@ -13,7 +13,9 @@ class Input extends Component {
   onSubmit(e) {
     e.preventDefault();
     this.setState({ text: "" });
-    this.props.onSendMessage(this.state.text);
+    if (this.state.text !== " ") {
+      this.props.onSendMessage(this.state.text);
+    }
   }
 
   render() {
